requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases
aiosqlite           # ← add this line
jinja2
python-multipart
bcrypt
python-dotenv
